refactor(dev-endpoints): extract networks-config leaf to break constants<->resolver cycle
